+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Braintree iOS SDK Release Notes
 
+## unreleased
+* BraintreeThreeDSecure
+  * Rename `AdditionalInformation` to `BTThreeDSecureAdditionalInformation`.
+  * Update `BTThreeDSecurePostalAddress` to make all properties accessible on the initializer only vs via the dot syntax.
+
 ## 7.0.0 (2025-11-04)
 * BraintreePayPal
   * Update `BTPayPalLineItem` to make all properties accessible on the initializer only vs via the dot syntax.